+++ core/src/main/scala/sangria/federation/v1/Decoder.scala
@@ -1,4 +1,4 @@
-package sangria.federation
+package sangria.federation.v1
 
 trait Decoder[Node, T] {
