+++ core/src/main/scala/sangria/federation/v1/Federation.scala
@@ -1,4 +1,4 @@
-package sangria.federation
+package sangria.federation.v1
 
 import sangria.ast.Document
 import sangria.marshalling.InputUnmarshaller
