+++ core/src/main/scala/sangria/federation/v1/NodeObject.scala
@@ -1,4 +1,4 @@
-package sangria.federation
+package sangria.federation.v1
 
 private[federation] trait NodeObject[Node] {
