+++ core/src/main/scala/sangria/federation/v1/_Entity.scala
@@ -1,4 +1,4 @@
-package sangria.federation
+package sangria.federation.v1
 
 import sangria.schema._
